split off document list
