+++ modules/frontend/src/app/documents/page.tsx
@@ -1,5 +1,5 @@
 
-import { DynamicConnected } from './components/dynamic'
+import { Dynamic } from './components/dynamic'
 
 import { Header } from '@/app/components/Header'
 
@@ -8,7 +8,7 @@ export default function Documents() {
   return (<div className={"container mx-auto"}>
      <Header hideExplaination={true} />
      <div className="m-2">
-        <DynamicConnected />
+        <Dynamic />
      </div>
   </div>)
 }
